refactor: start refactoring sgxvm
